server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
pdfplumber==0.11.1
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
sentence-transformers==2.7.0
numpy==1.26.4
scipy==1.13.0
transformers==4.40.2
torch==2.3.0
aiofiles==23.2.1
